8259778: Merge MutableSpace and ImmutableSpace
