+++ src/jdk.hotspot.agent/share/classes/sun/jvm/hotspot/gc/parallel/MutableSpace.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2003, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2003, 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -34,7 +34,7 @@ import sun.jvm.hotspot.types.*;
 import sun.jvm.hotspot.utilities.Observable;
 import sun.jvm.hotspot.utilities.Observer;
 
-public class MutableSpace extends ImmutableSpace {
+public class MutableSpace extends VMObject {
    static {
       VM.registerVMInitializedObserver(new Observer() {
          public void update(Observable o, Object data) {
@@ -45,6 +45,8 @@ public class MutableSpace extends ImmutableSpace {
 
    private static synchronized void initialize(TypeDataBase db) {
       Type type = db.lookupType("MutableSpace");
+      bottomField = type.getAddressField("_bottom");
+      endField    = type.getAddressField("_end");
       topField    = type.getAddressField("_top");
    }
 
@@ -53,14 +55,34 @@ public class MutableSpace extends ImmutableSpace {
    }
 
    // Fields
+   private static AddressField bottomField;
+   private static AddressField endField;
    private static AddressField topField;
 
    // Accessors
+   public Address   bottom()       { return bottomField.getValue(addr); }
+   public Address   end()          { return endField.getValue(addr);    }
    public Address   top()          { return topField.getValue(addr);    }
 
    /** In bytes */
    public long used() {
       return top().minus(bottom());
+   }
+
+   /** Returned value is in bytes */
+   public long capacity() { return end().minus(bottom()); }
+
+   /** Returns a subregion of the space containing all the objects in
+      the space. */
+   public MemRegion usedRegion() {
+      return new MemRegion(bottom(), end());
+   }
+
+   /** Support for iteration over heap -- not sure how this will
+      interact with GC in reflective system, but necessary for the
+      debugging mechanism */
+   public OopHandle bottomAsOopHandle() {
+      return bottomField.getOopHandle(addr);
    }
 
    /** returns all MemRegions where live objects are */
@@ -70,6 +92,12 @@ public class MutableSpace extends ImmutableSpace {
       return res;
    }
 
+   /** Testers */
+   public boolean contains(Address p) {
+      return (bottom().lessThanOrEqual(p) && end().greaterThan(p));
+   }
+
+   public void print() { printOn(System.out); }
    public void printOn(PrintStream tty) {
       tty.print(" [" + bottom() + "," +
                 top() + "," + end() + "] ");
